removed INT and INT64 types -> BIGINT; ima->IMA; (0,0)->CPX(0,0); other tidying from -Wall

+++ src/dirft.h
@@ -3,16 +3,16 @@
 
 #include "utils.h"
 
-void dirft1d1(INT nj,FLT* x,CPX* c,int isign,INT ms, CPX* f);
-void dirft1d2(INT nj,FLT* x,CPX* c,int iflag,INT ms, CPX* f);
-void dirft1d3(INT nj,FLT* x,CPX* c,int iflag,INT nk, FLT* s, CPX* f);
+void dirft1d1(BIGINT nj,FLT* x,CPX* c,int isign,BIGINT ms, CPX* f);
+void dirft1d2(BIGINT nj,FLT* x,CPX* c,int iflag,BIGINT ms, CPX* f);
+void dirft1d3(BIGINT nj,FLT* x,CPX* c,int iflag,BIGINT nk, FLT* s, CPX* f);
 
-void dirft2d1(INT nj,FLT* x,FLT *y,CPX* c,int iflag,INT ms, INT mt, CPX* f);
-void dirft2d2(INT nj,FLT* x,FLT *y,CPX* c,int iflag,INT ms, INT mt, CPX* f);
-void dirft2d3(INT nj,FLT* x,FLT *y,CPX* c,int iflag,INT nk, FLT* s, FLT* t, CPX* f);
+void dirft2d1(BIGINT nj,FLT* x,FLT *y,CPX* c,int iflag,BIGINT ms, BIGINT mt, CPX* f);
+void dirft2d2(BIGINT nj,FLT* x,FLT *y,CPX* c,int iflag,BIGINT ms, BIGINT mt, CPX* f);
+void dirft2d3(BIGINT nj,FLT* x,FLT *y,CPX* c,int iflag,BIGINT nk, FLT* s, FLT* t, CPX* f);
 
-void dirft3d1(INT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,INT ms, INT mt, INT mu, CPX* f);
-void dirft3d2(INT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,INT ms, INT mt, INT mu, CPX* f);
-void dirft3d3(INT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,INT nk, FLT* s, FLT* t, FLT *u, CPX* f);
+void dirft3d1(BIGINT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,BIGINT ms, BIGINT mt, BIGINT mu, CPX* f);
+void dirft3d2(BIGINT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,BIGINT ms, BIGINT mt, BIGINT mu, CPX* f);
+void dirft3d3(BIGINT nj,FLT* x,FLT *y,FLT *z,CPX* c,int iflag,BIGINT nk, FLT* s, FLT* t, FLT *u, CPX* f);
 
 #endif
